[frontend] Impove navigation to result page
Fix the error about destroied view by setting a timeout for redirection since some animation may not done.

Add detectchange in result page for display the data. However need 5 or more seconds to get the data properly displayed.
Seems the issue is how service are injected.No idea how to fix that so far.

Improved the appearence of result page.